requirements.txt
Django==2.1.7
gunicorn==19.9.0
psycopg2-binary==2.7.7
celery==4.2.1
redis==3.1.0
djangorestframework==3.9.1
django-oauth-toolkit==1.2.0
django-cors-middleware
